re-wrote arithmatic functions to take an int and an sexp * as parameters to fit my new way of doing things
ALSO: weirdest error ever.
I accidently defined type in debug.h, for some reason, which caused
some absolutely bizzare errors and took a while to figure out.

I also fixed a bunch of little things that casued complier errors, ie
things that aren't types anymore or enum values that had changed,etc..

+++ src/frames.h
@@ -48,5 +48,5 @@ void unwind_to_tag(env_ptr env,uint64_t tag) __attribute__((noreturn));
 void unwind_with_value(uint64_t tag,sexp value) __attribute__((noreturn));
 void unwind_call_stack(env_ptr env,uint64_t index);
 void unwind_bindings(env_ptr env,uint64_t n);
-#define raise_simple_error(tag,value) unwind_with_value(tag,value)
+#define raise_simple_error(tag,value) unwind_with_value((uint64_t)tag,value)
 #endif
